+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Added support for proven security estimation in `no_std` context (#218).
 * [BREAKING] refactored `verify()` function to take `AcceptableOptions` as a parameter (#219).
 * Increased min version of `rustc` to 1.73 (#221).
+* Allowed duplicate query positions (#224).
 
 ## 0.6.5 (2023-08-09) - math crate only
 * Added conditional support for serde on field elements (#209)